requirements.txt
# Python dependencies for the pipeline
PyYAML>=6.0
pandas>=1.5.0
numpy>=1.23.0
